requirements.txt
flask==1.1.2
flask_sqlalchemy==2.5.1
werkzeug==1.0.1
uuid==1.30
pyjwt==2.1.0